Added All_coins.py
This python script generalizes the augmentoai.py script to several coins at the same time. It retrieves the aggregated counts of topics mentioned for several coins then aggregate them into positive, neutral and negative topics for each of them. It also calculates the needed value to be taken into account for the Fear and Greed Index